Added indentation functionality
A good SSH config file is properly indented. While that was possible
to do manually, it's better if Emacs just does it for you. So I've
added a very crude implementation of the indentation. The premise is
this:

- if it's a comment or beginning of buffer, start at column 0
- if it's a line starting with 'Host', also indent to column 0
- all other SSH keywords should indent to 2 spaces

Now, I know that I've hardcoded '2 spaces', which should be
configurable, but for me it's what I'd like. And this is meant to be
just a demo of what it could be like. So before I go on a safari where
I try to add feature that are shiny and nice, I thought of throwing
this up and seeing if it's something you might be interested in.