Updated README.org
The README referred to 'ssh' (lower case) the SSH protocol.

The examples for autoload & auto-mode-alist are already being handled
by the mode, no need for it to be set out in the README.

Added an example of how to enable auto-mode-alist for a SSH config
file that isn't in an expected location.